[+] Add mileena configs creation from tpl.

See merge request cellframe/cellframe-node!471

+++ scripts/create_configs_from_tpl.sh
@@ -53,9 +53,9 @@ if [ "$DAP_CORE_T_ENABLED"="true" ]; then
     DAP_CFG_NET_TPL="$DAP_PREFIX/share/configs/network/$NET_NAME.cfg.tpl"
     DAP_NET_CFG=""
     if [ -e "$DAP_CFG_NET" ]; then
-	DAP_NET_CFG="$DAP_PREFIX/etc/network/$NET_NAME.cfg.new"
+    DAP_NET_CFG="$DAP_PREFIX/etc/network/$NET_NAME.cfg.new"
     else
-	DAP_NET_CFG="$DAP_PREFIX/etc/network/$NET_NAME.cfg"
+    DAP_NET_CFG="$DAP_PREFIX/etc/network/$NET_NAME.cfg"
     fi
 
     cat $DAP_CFG_NET_TPL > $DAP_NET_CFG || true
@@ -112,6 +112,23 @@ if [ "$DAP_KELVPN_MINKOWSKI_ENABLED"="true" ]; then
 
     cat $DAP_CFG_NET_TPL > $DAP_NET_CFG || true
     sed -i .old "s/{NODE_TYPE}/$DAP_KELVPN_MINKOWSKI_ROLE/" $DAP_NET_CFG  || true
+    rm $DAP_NET_CFG.old
+fi
+
+NET_NAME="mileena"
+
+if [ "$DAP_MILEENA_ENABLED"="true" ]; then
+    DAP_CFG_NET="$DAP_PREFIX/etc/network/$NET_NAME.cfg"
+    DAP_CFG_NET_TPL="$DAP_PREFIX/share/configs/network/$NET_NAME.cfg.tpl"
+    DAP_NET_CFG=""
+    if [ -e "$DAP_CFG_NET" ]; then
+    DAP_NET_CFG="$DAP_PREFIX/etc/network/$NET_NAME.cfg.new"
+    else
+    DAP_NET_CFG="$DAP_PREFIX/etc/network/$NET_NAME.cfg"
+    fi
+
+    cat $DAP_CFG_NET_TPL > $DAP_NET_CFG || true
+    sed -i .old "s/{NODE_TYPE}/$DAP_MILEENA_ROLE/" $DAP_NET_CFG  || true
     rm $DAP_NET_CFG.old
 fi
